Style: rename common files to common idiom names
- Terms like 'filesystem' and 'timestamp' are common-place and do not
  need an underscore
- Reverted diffie-hellman to diffie_hellman per our upstream style guide
  for consistency
