+++ src/client/address_book.h
@@ -45,7 +45,7 @@
 #include <vector>
 
 #include "identity.h"
-#include "core/util/file_system.h"
+#include "core/util/filesystem.h"
 #include "util/base64.h"
 #include "util/log.h"
