+++ src/client/i2p_control/i2p_control_server.cc
@@ -41,7 +41,7 @@
 
 #include "core/version.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace client {
