+++ src/client/streaming.cc
@@ -41,7 +41,7 @@
 #include "crypto/util/compression.h"
 #include "tunnel/tunnel.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace stream {
