+++ src/core/i2np_protocol.cc
@@ -49,7 +49,7 @@
 #include "tunnel/tunnel.h"
 #include "util/i2p_endian.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 #ifndef NETWORK_ID
 #define NETWORK_ID 2
