+++ src/core/router_info.cc
@@ -46,7 +46,7 @@
 #include "util/base64.h"
 #include "util/i2p_endian.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace data {
