+++ src/core/transport/ssu.cc
@@ -47,7 +47,7 @@
 #include "net_db.h"
 #include "router_context.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace transport {
