+++ src/core/transport/ssu_data.cc
@@ -42,7 +42,7 @@
 #include "net_db.h"
 #include "ssu.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace transport {
