+++ src/core/transport/ssu_session.cc
@@ -41,11 +41,11 @@
 #include "router_context.h"
 #include "ssu.h"
 #include "transports.h"
-#include "crypto/diffie-hellman.h"
+#include "crypto/diffie_hellman.h"
 #include "crypto/hash.h"
 #include "crypto/rand.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace transport {
