+++ src/core/tunnel/tunnel.cc
@@ -48,7 +48,7 @@
 #include "transport/transports.h"
 #include "util/i2p_endian.h"
 #include "util/log.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace tunnel {
