+++ src/core/tunnel/tunnel_config.cc
@@ -37,7 +37,7 @@
 #include "crypto/rand.h"
 #include "i2np_protocol.h"
 #include "router_context.h"
-#include "util/time_stamp.h"
+#include "util/timestamp.h"
 
 namespace i2p {
 namespace tunnel {
